1683-update-heat-agent-queries: Added methods to calculate heat emission rates of chemical plants in mainland Singapore.
